Fix platform-specific framerate & text render issues
On a Windows machine, font.render() with no background parameter was producing a black background, while it was producing a transparent background on a MacOS machine.
I believe the latter is the expected behavior in Pygame 2.0.0.dev6, but I went ahead and hardcoded the beige color under the textbox as its font.render() background.
Also added a 30 FPS clock to deal with various processor speeds and text scrolling.